tests/wpt/web-platform-tests/content-security-policy/unsafe-hashes/style_attribute_allowed.html

<!DOCTYPE HTML>
<html>

<head>
    <meta http-equiv="Content-Security-Policy" content="img-src *;
                      style-src 'unsafe-hashes' 'sha256-S0VSqEOmzmyOifPfat2sJ7ELOgkldAEbaXlvi5iMqjc=';">
    <!--
      'sha256-S0VSqEOmzmyOifPfat2sJ7ELOgkldAEbaXlvi5iMqjc=' ==> 'background: green'
      -->
    <title>Event handlers should be allowed if a matching hash and 'unsafe-hashes' are present</title>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
</head>

<body>
    <div id='log'></div>
    <script>
        var t1 = async_test("Test that the inline style attribute is loaded");

        function check_for_style() {
          assert_equals("green", document.getElementById('test').style.background);
          t1.done();
        }

        window.addEventListener('securitypolicyviolation', t1.unreached_func("Should have not raised any event"));
    </script>
    <img src='../support/pass.png' id='test' style='background: green'
         onload='check_for_style()'>
</body>

</html>
