Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/unsafe-hashes/style_attribute_allowed-expected.txt


FAIL Test that the inline style attribute is loaded assert_equals: expected "" but got "green"
